moved test for each crawler to its corresponding file
