+++ tests/scraper/crawler/primicia_crawler_test.py
@@ -1,0 +1,65 @@
+from c4v.scraper.crawler.crawlers.primicia_crawler import PrimiciaCrawler
+
+
+def test_crawler_parse_primicia_index():
+    prm = PrimiciaCrawler()
+
+    out = prm.parse_sitemaps_urls_from_index(primicia_sitemap_index())
+    ans = ['https://primicia.com.ve/post-sitemap1.xml', 'https://primicia.com.ve/post-sitemap2.xml', 'https://primicia.com.ve/post-sitemap3.xml']
+
+    assert set(out) == set(ans)
+
+def test_crawler_parse_primicia_sitemap():
+    prm = PrimiciaCrawler()
+
+    out = prm.parse_urls_from_sitemap(primicia_sitemap_body())
+    ans = ["https://primicia.com.ve/sucesos/un-muerto-y-cuatro-heridos-durante-protestas/", "https://primicia.com.ve/mas/salud/the-lancet-global-health-aumentan-muertes-infantiles-en-venezuela/"]
+    
+    assert set(out) == set(ans)
+
+def primicia_sitemap_index():
+    return '<?xml version="1.0" encoding="UTF-8"?><?xml-stylesheet type="text/xsl" href="//primicia.com.ve/wp-content/plugins/wordpress-seo/css/main-sitemap.xsl"?>\
+<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">\
+        <sitemap>\
+                <loc>https://primicia.com.ve/post-sitemap1.xml</loc>\
+                <lastmod>2019-02-15T01:27:26+00:00</lastmod>\
+        </sitemap>\
+        <sitemap>\
+                <loc>https://primicia.com.ve/post-sitemap2.xml</loc>\
+                <lastmod>2019-03-07T19:03:35+00:00</lastmod>\
+        </sitemap>\
+        <sitemap>\
+                <loc>https://primicia.com.ve/post-sitemap3.xml</loc>\
+                <lastmod>2019-03-26T04:33:30+00:00</lastmod>\
+        </sitemap>\
+        <sitemap>\
+                <loc>https://primicia.com.ve/author-sitemap.xml</loc>\
+                <lastmod>2020-04-28T17:47:34+00:00</lastmod>\
+        </sitemap>\
+        </sitemapindex>'
+
+def primicia_sitemap_body():
+    return '<?xml version="1.0" encoding="UTF-8"?><?xml-stylesheet type="text/xsl" href="//primicia.com.ve/wp-content/plugins/wordpress-seo/css/main-sitemap.xsl"?>\
+<urlset xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd http://www.google.com/schemas/sitemap-image/1.1 http://www.google.com/schemas/sitemap-image/1.1/sitemap-image.xsd" xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">\
+	<url>\
+		<loc>https://primicia.com.ve/</loc>\
+		<lastmod>2021-05-21T22:15:27+00:00</lastmod>\
+	</url>\
+	<url>\
+		<loc>https://primicia.com.ve/sucesos/un-muerto-y-cuatro-heridos-durante-protestas/</loc>\
+		<lastmod>2019-01-25T15:30:32+00:00</lastmod>\
+		<image:image>\
+			<image:loc>https://primicia.com.ve/wp-content/uploads/2019/01/protestaas.jpg</image:loc>\
+			<image:title><![CDATA[protestaas]]></image:title>\
+		</image:image>\
+	</url>\
+	<url>\
+		<loc>https://primicia.com.ve/mas/salud/the-lancet-global-health-aumentan-muertes-infantiles-en-venezuela/</loc>\
+		<lastmod>2019-01-25T15:35:34+00:00</lastmod>\
+		<image:image>\
+			<image:loc>https://primicia.com.ve/wp-content/uploads/2019/01/hospital.jpg</image:loc>\
+			<image:title><![CDATA[hospital]]></image:title>\
+		</image:image>\
+	</url>\
+</urlset>\
+<!-- XML Sitemap generated by Yoast SEO -->'
